LPS-61133 move packages
