+++ modules/portal-security/portal-security-sso-facebook-connect/src/main/resources/META-INF/resources/com.liferay.portal.settings.web/init.jsp
@@ -23,8 +23,8 @@ taglib uri="http://liferay.com/tld/theme" prefix="liferay-theme" %>
 page import="com.liferay.portal.kernel.settings.CompanyServiceSettingsLocator" %><%@
 page import="com.liferay.portal.kernel.settings.ParameterMapSettingsLocator" %><%@
 page import="com.liferay.portal.kernel.util.Validator" %><%@
+page import="com.liferay.portal.security.sso.facebook.connect.configuration.FacebookConnectConfiguration" %><%@
 page import="com.liferay.portal.security.sso.facebook.connect.constants.FacebookConnectConstants" %><%@
-page import="com.liferay.portal.security.sso.facebook.connect.module.configuration.FacebookConnectConfiguration" %><%@
 page import="com.liferay.portal.util.Portal" %>
 
 <%@ page import="javax.portlet.ActionRequest" %>
